% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{addDefaultArgs}
\alias{addDefaultArgs}
\title{Add one element into a list of arguments}
\usage{
addDefaultArgs(defaultArgs, ui_element, ui_param, plot_settings)
}
\arguments{
\item{defaultArgs}{the default argument list}

\item{ui_element}{the keyword of the ui element}

\item{ui_param}{the parameter of the ui function}

\item{plot_settings}{a list contains all ui elements}
}
\value{
a list of default arguments
}
\description{
Add one element into a list of arguments
}
