% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_server_handler.R
\name{facetHandler}
\alias{facetHandler}
\title{Module server for \code{facet_grid()}}
\usage{
facetHandler(id, module_id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{module_id}{the id of the ui element (not including its prefix created by
the name space) which gives input to \code{facet_grid()}}
}
\value{
\code{facet_grid()} and its code or \code{NULL}
}
\description{
Module server for \code{facet_grid()}
}
