% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr.R
\name{get_plot}
\alias{get_plot}
\title{Extract plot}
\usage{
get_plot(data, gg_list)
}
\arguments{
\item{data}{data used for the plot}

\item{gg_list}{list of \code{gg} objects}
}
\value{
\code{ggplot} object
}
\description{
Extract plot
}
