% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr.R
\name{get_plot_code}
\alias{get_plot_code}
\title{Extract code and plot}
\usage{
get_plot_code(componentList, data, data_path, data_filter = "")
}
\arguments{
\item{componentList}{list of plot components and code components}

\item{data}{data}

\item{data_path}{path to the data}

\item{data_filter}{data filtering code}
}
\value{
a named list of two elements; plot and code
}
\description{
Extract code and plot
}
