% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_control.R
\name{matchControls}
\alias{matchControls}
\title{match key to its ui or handler function}
\usage{
matchControls(selected, scope = "mapping", type = "ui")
}
\arguments{
\item{selected}{the key to be matched}

\item{scope}{optional. \code{scope} can be one of \code{mapping}, \code{geom_args}, or \code{plot_settings}}

\item{type}{optional. By default it's \code{ui}. Can be \code{ui} or \code{handler}}
}
\value{
the ui or handler function of \code{selected} key
}
\description{
match key to its ui or handler function
}
