% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{run_ggpaintr_app}
\alias{run_ggpaintr_app}
\title{Call ggpaintr_app}
\usage{
run_ggpaintr_app(name = NULL)
}
\arguments{
\item{name}{name of the installed shiny app. Default is \code{NULL}. Other choices
include: \code{ggpaintr_demo}}
}
\value{
\code{shiny::runApp}
}
\description{
Call ggpaintr_app
}
