% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_server_handler.R
\name{scaleColorFillHandler}
\alias{scaleColorFillHandler}
\title{Module server that sets colors through \code{scale_fill_gradient}, \code{scale_fill_manual},
\code{scale_color_gradient}, \code{scale_color_manual}}
\usage{
scaleColorFillHandler(id, selected_color_fill_rctv, color_fill)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{selected_color_fill_rctv}{the reactive values returned by \code{scaleColor_build_reactivity()}}

\item{color_fill}{string. either \code{color} or \code{fill}. Specifies whether it's \code{scale_color}
or \code{scale_fill}}
}
\value{
the return of \verb{scale_<fill/color>_<gradient/manual>} and its code or \code{NULL}
}
\description{
Module server that sets colors through \code{scale_fill_gradient}, \code{scale_fill_manual},
\code{scale_color_gradient}, \code{scale_color_manual}
}
\note{
this function handles the \code{error} when \code{selected_color_fill_rctv()} returns error
}
