% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{unwrap_expr}
\alias{unwrap_expr}
\title{Unwrap an expression}
\usage{
unwrap_expr(x)
}
\arguments{
\item{x}{an expression}
}
\value{
list of strings converted from the expression
}
\description{
Unwrap an expression
}
\examples{
unwrap_expr(x + y + z)
}
